first commit -> added BigDecimal

+++ kodilla-stream/src/main/java/com/kodilla/stream/sand/SandStorage.java
@@ -1,5 +1,7 @@
 package com.kodilla.stream.sand;
 
+import java.math.BigDecimal;
+
 public interface SandStorage {
-    long getSandBeansQuantity();
+    BigDecimal getSandBeansQuantity();
 }
